vertex-pipeline/pyproject.toml
[project]
name = "vertex-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-pipeline-components>=2.22.0",
    "numpy<2.0.0",
    "pandas==2.2.2"
]
